requirements.txt
django_compressor==2.2
django-clear-cache==0.3
django-debug-toolbar==1.9.1
django-memoize==2.1.0
django-sass-processor==0.5.5
django==1.11.6
gunicorn==19.7.1
libsass==0.13.3
markdown==2.6.9
pillow==4.3.0
pygments==2.2.0
pyquery==1.3.0
python-frontmatter==0.2.1
setproctitle==1.1.10
